Декомпозиция Profile
