+++ src/pages/Profile/Profile.tsx
@@ -1,70 +1,25 @@
 import { Button, Form, Col, Row, Input, Space, Divider } from "antd";
-import { message, Upload, Avatar, Flex } from "antd";
+import { message, Avatar, Flex } from "antd";
 import { Typography } from "antd";
 import TextArea from "antd/es/input/TextArea";
-import { UploadChangeParam } from "antd/es/upload";
-import { RcFile, UploadFile, UploadProps } from "antd/es/upload/interface";
 
 import { useEffect, useReducer, useState } from "react";
 import { useParams } from "react-router-dom";
 
-import { UploadOutlined } from "@ant-design/icons";
-import {
-  ProfileInfoDto,
-  UpdateProfileError,
-  userApi
-} from "@api/user/user.api";
+import { UpdateProfileError, userApi } from "@api/user/user.api";
 import { EditableFormItem } from "@components/EditableFormItem/EditableFormItem";
 import { InfoIcon } from "@components/Icons/InfoIcon";
 import { Loader } from "@components/Loader/Loader";
 import { TagPool } from "@components/TagPool/TagPool";
-import { getBase64, beforeUpload } from "@infrastructure/image-upload";
 import { Role } from "@interfaces/user";
 
+import { EditableHeader } from "./Profile.Header";
 import styles from "./Profile.module.scss";
+import { IProfileCompound } from "./Profile.Types";
+import { ProfileReducer } from "./ProfileReductor";
 
-type ProfileReducerAction = {
-  type: "set_user" | "change_last_name" | "change_first_name"; // eslint-disable-next-line
-  value?: any;
-};
-function reducer(
-  state: ProfileInfoDto | null,
-  action: ProfileReducerAction
-): ProfileInfoDto | null {
-  let data = null;
-  if (state !== null) data = { ...state, user: { ...state.user } };
-  switch (action.type) {
-    case "set_user": {
-      if (action.value === undefined)
-        throw Error(
-          `action.value must be not null for action.type === ${action.type}`
-        );
-      data = { ...action.value };
-      return data;
-    }
-    case "change_first_name": {
-      if (action.value === undefined)
-        throw Error(
-          `action.value must be not null for action.type === ${action.type}`
-        );
-      if (data === null) throw Error(`state must be not null`);
-      data.user.first_name = action.value;
-      return data;
-    }
-    case "change_last_name": {
-      if (action.value === undefined)
-        throw Error(
-          `action.value must be not null for action.type === ${action.type}`
-        );
-      if (data === null) throw Error(`state must be not null`);
-      data.user.last_name = action.value;
-      return data;
-    }
-  }
-}
-
-export const Profile = () => {
-  const [profileData, dispatch] = useReducer(reducer, null);
+export const Profile: React.FC & IProfileCompound = () => {
+  const [profileData, dispatch] = useReducer(ProfileReducer, null);
   const { profileId } = useParams();
   const [isLoading, setIsLoading] = useState(false);
   const [isReadOnly, setIsReadOnly] = useState(true);
@@ -93,28 +48,6 @@ export const Profile = () => {
       setIsReadOnly(true);
     }
   }, [profileId]);
-
-  const handleChange: UploadProps["onChange"] = (
-    info: UploadChangeParam<UploadFile>
-  ) => {
-    if (info.file.status === "uploading") {
-      setIsLoading(true);
-    } else if (info.file.status === "done") {
-      getBase64(info.file.originFileObj as RcFile, (url) => {
-        setImageUrl(url);
-        setIsLoading(false);
-      });
-    } else {
-      setIsLoading(false);
-    }
-  };
-  // eslint-disable-next-line
-  const dummyRequest = (options: any) => {
-    // Пока что сделать заглушку для upload фото на сервер
-    setTimeout(() => {
-      options.onSuccess("ok");
-    }, 0);
-  };
 
   const handleSave = () => {
     if (!profileData) return;
@@ -145,11 +78,7 @@ export const Profile = () => {
   return (
     <>
       <Loader active={isLoading} />
-      {!profileData ? (
-        <div style={{ padding: "1rem" }}>
-          <Typography.Title>Пользователь не найден.</Typography.Title>
-        </div>
-      ) : (
+      {profileData && (
         <Form
           layout={"vertical"}
           initialValues={{ layout: "horizontal" }}
@@ -175,86 +104,15 @@ export const Profile = () => {
                     readonly={isReadOnly}
                   >
                     <EditableFormItem.EditablePart>
-                      <Row gutter={[12, 12]}>
-                        <Col span={8}>
-                          <Form.Item label="Имя">
-                            <Input
-                              value={profileData?.user.first_name}
-                              onChange={(e) =>
-                                dispatch({
-                                  type: "change_first_name",
-                                  value: e.target.value
-                                })
-                              }
-                              readOnly={isReadOnly}
-                            />
-                          </Form.Item>
-                          <Form.Item label="Фамилия">
-                            <Input
-                              value={profileData?.user.last_name}
-                              onChange={(e) =>
-                                dispatch({
-                                  type: "change_last_name",
-                                  value: e.target.value
-                                })
-                              }
-                              readOnly={isReadOnly}
-                            />
-                          </Form.Item>
-                          <Form.Item label="Telegram">
-                            <Input
-                              placeholder="Soon..."
-                              readOnly={isReadOnly}
-                            />
-                          </Form.Item>
-                        </Col>
-                        <Col span={8}>
-                          <Form.Item label="Почта">
-                            <Input
-                              value={profileData?.user.email}
-                              readOnly={isReadOnly}
-                            />
-                          </Form.Item>
-                          <Form.Item label="Номер телефона">
-                            <Input
-                              placeholder="Soon..."
-                              readOnly={isReadOnly}
-                            />
-                          </Form.Item>
-                        </Col>
-                        <Col span={8}>
-                          {!isReadOnly && (
-                            <Flex
-                              style={{ width: "100%" }}
-                              justify={"end"}
-                              align={"end"}
-                              gap={10}
-                            >
-                              <Upload
-                                showUploadList={false}
-                                customRequest={dummyRequest}
-                                beforeUpload={beforeUpload}
-                                onChange={handleChange}
-                              >
-                                <Button
-                                  icon={<UploadOutlined />}
-                                  type="primary"
-                                  size="large"
-                                >
-                                  Добавить фото
-                                </Button>
-                              </Upload>
-                              <Button
-                                danger
-                                style={{ backgroundColor: "white" }}
-                                size="large"
-                              >
-                                Удалить
-                              </Button>
-                            </Flex>
-                          )}
-                        </Col>
-                      </Row>
+                      <Profile.HeaderEditable
+                        dispatch={dispatch}
+                        email={profileData?.user.email}
+                        first_name={profileData?.user.first_name}
+                        last_name={profileData?.user.last_name}
+                        isReadOnly={isReadOnly}
+                        setImageUrl={setImageUrl}
+                        setIsLoading={setIsLoading}
+                      />
                     </EditableFormItem.EditablePart>
                     <EditableFormItem.ReadOnlyPart>
                       <Row>
@@ -289,6 +147,7 @@ export const Profile = () => {
             </Col>
           </Row>
           <Divider style={{ borderColor: "#7E7E7E66" }} />
+
           <Row>
             <Col span={24}>
               <Form.Item label="Должность">
@@ -325,3 +184,5 @@ export const Profile = () => {
     </>
   );
 };
+
+Profile.HeaderEditable = EditableHeader;
